+++ app/src/main/java/com/example/myapplication/model/models/Album.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.models
+package com.example.myapplication.model.models
 
 import org.json.JSONObject
