+++ app/src/main/java/com/example/myapplication/view/CollectorFragment.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.ui
+package com.example.myapplication.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,7 +8,7 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.example.myapplication.databinding.FragmentCollectorsBinding
-import com.example.myapplication.viewmodels.CollectorViewModel
+import com.example.myapplication.viewModel.CollectorViewModel
 
 class CollectorFragment : Fragment() {
