+++ app/src/main/java/com/example/myapplication/viewModel/CollectorViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.viewmodels
+package com.example.myapplication.viewModel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
